+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added Spanish summarisation dataset DACSA-es as an unofficial dataset.
 - Added Lithuanian sentiment classification dataset Atsiliepimai to replace the now
   unofficial Lithuanian Emotions dataset.
 - Added support for Hungarian 🇭🇺! This includes the sentiment classification dataset